+++ src/main/java/io/contek/invoker/mexc/spot/api/common/_Balance.java
@@ -1,7 +1,7 @@
 package io.contek.invoker.mexc.spot.api.common;
 
 import javax.annotation.concurrent.NotThreadSafe;
-import java.math.BigDecimal;
+import java.math.Double;
 
 @NotThreadSafe
 public class _Balance {
@@ -10,6 +10,6 @@ public class _Balance {
   //  //        "locked": "33"
 
   public String asset;
-  public BigDecimal free;
-  public BigDecimal locked;
+  public Double free;
+  public Double locked;
 }
