+++ src/main/java/io/contek/invoker/mexc/spot/api/common/_Ticker.java
@@ -1,17 +1,17 @@
 package io.contek.invoker.mexc.spot.api.common;
 
 import javax.annotation.concurrent.NotThreadSafe;
-import java.math.BigDecimal;
+import java.math.Double;
 
 @NotThreadSafe
 public class _Ticker {
 
   public String symbol;
 
-  public BigDecimal askQty;
-  public BigDecimal askPrice;
-  public BigDecimal bidQty;
-  public BigDecimal bidPrice;
+  public Double askQty;
+  public Double askPrice;
+  public Double bidQty;
+  public Double bidPrice;
 
   public long traceNano = System.nanoTime();
 }
